dashboard: make setup_pipeline funcs idempotent
When upgrading the qa environments using "make all", a new pipeline is
created on every upgrade and registered on the SS when setup_pipeline
and setup_pipeline_in_ss are invoked in steps.py. This change will make
both methods idempotent, so that:

- if the pipeline already exists no new uuid will be assigned to the
existing pipeline.

- if the pipeline is registered on the SS, no further action will be
needed.

The _get_pipeline function is renamed to get_pipeline because it is
used by other modules.

This fixes Jisc issues:
- JiscRDSS/rdss-archivematica#134
- JiscRDSS/rdss-archivematica#165